General interpolation for discrete fields and GPs
A new interpolation mechanism for discrete fields is introduced.
Compared to the original implementation, the new method takes an
interpolator object with the interpolation algorithm. This makes it possible
to use different interoplation algorithms using the same interface and separates
the interpolation algorithm from the field implementation. The same mechanism is also
used for the interpolation of a DiscreteLowRankGaussianProcess.