WIP Introduced PointSet as new abstraction.
The concept of PointSet was introduce, which essentially is a renaiming
of the DiscreteDomain. DiscreteDomain instead now only supports two
operations: To return the pointset and to return the topology. This
makes it possible to make the meshes a Domain, and hence using them
directly as a domain for discretefield and discreteGaussianProcess